Проверка № 2

+++ src/main/java/ru/kata/spring/boot_security/demo/service/UserService.java
@@ -12,9 +12,9 @@ public interface UserService {
 
     void saveUsers(User user);
 
-    User getUser(int id);
+    User getUser(long id);
 
     void updateUser(User user);
 
-    void deleteUser(int id);
+    void deleteUser(long id);
 }
